moved plotting functions
