+++ src/server/nautical/synthtest/BoatSimDemo.cpp
@@ -4,7 +4,6 @@
  */
 
 #include <server/nautical/synthtest/BoatSim.h>
-#include <server/plot/extra.h>
 #include <server/common/string.h>
 #include <iostream>
 
@@ -18,53 +17,13 @@ namespace {
       };
   }
 
-  Arrayd getTimes(Array<BoatSimulator::FullBoatState> states) {
-    return states.map<double>([=](BoatSimulator::FullBoatState s) {
-      return s.time.seconds();
-    });
-  }
 
-
-  void plotAngles(const char *title, Array<BoatSimulator::FullBoatState> states,
-      std::function<Angle<double>(BoatSimulator::FullBoatState)> fun) {
-    GnuplotExtra plot;
-    plot.set_title(title);
-    plot.set_style("lines");
-    plot.plot_xy(getTimes(states),
-        states.map<Angle<double> >(fun).map<double>([&](Angle<double> x)
-            {return x.degrees();}));
-    plot.show();
-  }
-
-  void plotSpeed(const char *title, Array<BoatSimulator::FullBoatState> states,
-      std::function<Velocity<double>(BoatSimulator::FullBoatState)> fun) {
-    GnuplotExtra plot;
-    plot.set_title(title);
-    plot.set_style("lines");
-    plot.plot_xy(getTimes(states),
-        states.map<Velocity<double> >(fun).map<double>([&](Velocity<double> x)
-            {return x.knots();}));
-    plot.show();
-  }
-
-  void plotTrajectory(Array<BoatSimulator::FullBoatState> states) {
-    GnuplotExtra plot;
-    plot.set_title("Trajectory (meters)");
-    plot.set_style("lines");
-    plot.plot_xy(
-        states.map<double>([&](BoatSimulator::FullBoatState x) {return x.x.meters();}),
-        states.map<double>([&](BoatSimulator::FullBoatState x) {return x.y.meters();})
-    );
-    plot.show();
-  }
 }
 
 int main(int argc, const char **argv) {
   Duration<double> dur = Duration<double>::minutes(0.5);
 
   BoatCharacteristics ch;
-
-
   auto windfun = makeConstantFlow(Velocity<double>::metersPerSecond(8),
                                   Angle<double>::degrees(0));
   auto currentfun = makeConstantFlow(Velocity<double>::knots(0.1),
@@ -79,11 +38,7 @@ int main(int argc, const char **argv) {
   Array<BoatSimulator::FullBoatState> states = simulator.simulate(1.99*dur,
       Duration<double>::seconds(0.05), 1);
 
-
-  plotAngles("Rudder angle (degrees)", states, [](const BoatSimulator::FullBoatState &x) {return x.rudderAngle;});
-  plotAngles("Boat orientation (degrees)", states, [](const BoatSimulator::FullBoatState &x) {return x.boatOrientation;});
-  plotSpeed("Boat speed through water (knots)", states, [](const BoatSimulator::FullBoatState &x) {return x.boatSpeedThroughWater;});
-  plotTrajectory(states);
+  BoatSimulator::makePlots(states);
   return 0;
 }
